refactored the code to be a little more readable, and to be compatible with original code.
